Reduce UART Thread Stack Size
The UART thread stack size was set to 4096, the same as the stack size
for the thread running TouchGFX.  It would appear using a stack this
large exhausts the space for the FreeRTOS heap (or some other pool of
storage for thread stacks), and thus the call to osThreadCreate was
returning NULL (i.e. the thread wasn't getting created).

This change fixes the issue by reducing the UART tasks's stack size.
The size was chosen using empirical observation.  Stack sizes of 128 and
256 would allow the thread to be created, but a HardFault would occur on
task switching, a sure sign of stack exhaustion.  At 512, there is
enough room for the standard debug loop to operate.  This may need to
grow as we go.

Note that this isn't really the right way to set up threads using an
STM32CubeIDE project manager.  In fact, the correct way to create a new
task is to open the .ioc file and go to Middleware->FreeRTOS->"Tasks and
Queues" and create one there.  The SMT32CubeIDE will manage the code and
also track how much of the FreeRTOS heap has been used.  This will be
corrected in a future commit.