Define UARTTask using STM32CubeIDE Tools
This commit changes the definition of the UARTTask so that it goes
through the STM32CubeIDE tools rather than being done by hand.  The
effective code is the same, but doing it this way lets us lean on the
IDE tools to help spot potential problems, like FreeRTOS heap
exhaustion, in the future.  Also, it ensures generated code is protected
from future runs of the code generator.